Patch 3 for shutdowns

+++ UserTools/SC_Emergency/SC_Emergency.cpp
@@ -98,7 +98,7 @@ bool SC_Emergency::HVCHK()
 
     if(m_data->SCMonitor.HV_return_mon < (m_data->SCMonitor.HV_volts-200) || m_data->SCMonitor.HV_return_mon > (m_data->SCMonitor.HV_volts+200))
     {
-        int down_voltage = 0;
+        float down_voltage = 0;
         m_data->SCMonitor.errorcodes.push_back(0xCC10EE01);
         m_data->CB->SetHV_voltage(down_voltage,m_data->SCMonitor.HV_return_mon,2);
